blobtools/pyproject.toml
[tool.poetry]
name = "blobtools"
version = "0.1.0"
description = ""
authors = ["Andre Bossard <andre.bossard@microsoft.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
opentelemetry-sdk = "^1.21.0"
azure-monitor-opentelemetry-exporter = "^1.0.0b19"
azure-storage-blob = "^12.19.0"
azure-storage-queue = "^12.9.0"
azure-identity = "^1.15.0"
azure-core = "^1.29.5"
python-dotenv = "^1.0.0"
opentelemetry-exporter-zipkin = "^1.21.0"
pydantic = "^2.5.2"
azure-core-tracing-opentelemetry = "^1.0.0b11"
aiohttp = "^3.9.1"
azure-monitor-opentelemetry = "^1.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
